pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pathway-explorer"
version = "2.0.0"
description = "Interactive GSEA/TF/TE pathway explorer dashboard generator"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"

authors = [
    {name = "SciAgent", email = "noreply@sciagent.dev"}
]

keywords = [
    "bioinformatics",
    "GSEA",
    "pathway analysis",
    "visualization",
    "RNA-seq",
    "transcription factors",
    "transposable elements"
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "scipy>=1.9.0",
]

[project.optional-dependencies]
full = [
    "umap-learn>=0.5.0",
    "scikit-learn>=1.0.0",
    "pyyaml>=6.0",
]

dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.10.0",
]

[project.scripts]
pathway-explorer = "pathway_explorer.main:main"

[project.urls]
Homepage = "https://github.com/your-org/pathway-explorer"
Documentation = "https://github.com/your-org/pathway-explorer#readme"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100
